+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 - Add MPM beam twist, snow ball, and viscous coiling examples
 - Add support for textures in `SensorTiledCamera` via `Config.enable_textures`
 - Add `enable_ambient_lighting` and `enable_particles` options to `SensorTiledCamera.Config`
-- Add differentiable rigid contacts with respect to body poses via `CollisionPipeline` when `requires_grad=True`
+- Add differentiable rigid contacts (experimental) with respect to body poses via `CollisionPipeline` when `requires_grad=True`
 
 ### Changed